requirements.txt
fastapi
uvicorn
redis
rq
httpx
python-dotenv